Removes excessive insanity in TGUI
